Docs: update introduction with signing code
All commits, merges, and pull requests must be signed by the author's PGP key. This ensures integrity of the source code comes from the person it says it does. It also allows use to 100% identify who is responsible for the code and any ethic violations. Stating it up front on the introduction screen ensures it is one of the first thing someone will see.